A3/requirements.txt
torch
torchvision
pandas
Pillow
sklearn
seaborn
matplotlib
numpy